Implement move to folder per D&D
We have to trigger the drop explicitly in onReleased.

Interestingly if we use "Drag.dragType: Drag.Automatic" then we don't
need to explicitly call drop() and the drag and drop behaviour works as
expected, but we end up no longer dragging the actual item from the
listview, which is what we wanted to do (although we'll have to deal
with the item reappering in the list only to then vanish from the actual
move).

It seems Drag.Automatic triggers a new style of drag that is now
recommended but entirely undocumented, and we're doing some old-style/
backwardscompatible drag...