Avoid "Unable to assign...." errors
i.e. "Unable to assign [undefined] to Sink::ApplicationDomain::Folder::Ptr"
Because qml does not know how to assign a default-constructed value we
have to use conditional bindings. The !! operator checks for
availability of the property which then triggers the binding.

I assume an alternative viable approach would be to expose all
controller properties as QVariants only, in which case an assignment
from null would work (it doesn't for Folder::Ptr).