Mark the currently viewed mail as read.
The listview is absolutely useless with non-uniform entries.
We have to do or own tracking of the current index, we have to do our
own scrolling, the flickable scrolling is horrible with a mouse,
and even though caching all delegates somewhat improves the situation,
it's still laggy every now and then.

However, this now at least works somewhat correctly.
But we'll really have to look for a new solution for this eventually.