Consistently give focus to popups
In all cases we want the popup to receive keyboard focus when
it's open, as all our popups are basically modal. We also want the popup
to normally close when clicking somewhere else, including the parent
button. The one exception is the composer where we disable the close on
click outside because that would risk loosing a written mail on an
accidental click.