Fixed resizing issues
Properly anchor all components and then let the layout deal with it.
Setting a maximum size for splits that depends on the application window
breaks initial layouting, probably because the window size initially is
0 and that results in the default size not being applied.
Since I don't think the maximum size is overly useful we can live
without for the time being.