Use a ScrollHelper to fix scrolling on listviews
Because the standard scrolling is so unusable depending on the input
device we replace it by something custom that is fairly similar to what
QQC1 ScrollView did. Using a ScrollView sucks in many ways, including
that you have to wrap all sorts of things which is just code wise not
great at all. The ScrollHelper can instead be attached to any existing
flickable to override it's scrolling behaviour, so we can also silently
drop it once the default flickable behaviour starts to make sense.