Put the TreeView into an extra flickable
...so it becomes scrollable just like the rest.
Because the TreeView derives from ScrollView we can't just
use the regular approach of making the flickable suck less at scrolling.
Instead we expand the treeview to maximum length and wrap it in an extra
Flickable that we can then handle with the ScrollHelper.
This results in the treeview to have the same scrolling behvaiour as
everything else.