Only highlight on right-click
With the highlight on hover the items look clickable, but they are not.
Further The UI starts to blink if you run the curser accross a couple of
labels. The new behaviour is more in line with what i.e. browsers do
with links or alike.

The highglight was also turned into a softer full overlay instead of
just a border.