Fixed mouse interaction with mail content
Previously mouse interaction was blocked by the mouse area in the
delegate. An additional problem was after that, that the listview lost
focus when the textview got focus which was fixed by introducing a
focus scope and aquiring the currentItem status via activeFocus.

The downside is that the focus stealing of the webengineview now becomes
apparent again, but that can be fixed in qt 5.9.
Otherwise this seems to work as expected now.