Fixed controller clear function
Adding a property from QML changes the offset from 1 to 20 for the
composercontroller, which results in the clear function not working.
Since we know that QObjects property offset is 1, we can just hardcode
that number. Not the cleanest solution, so it would be good to find a
better approach at some point.