Fixed initial account login.
Because we ended up trying to load the resources before they even
existed (directly after the account was created), we ended up creating a
second set of unconfigured resources.
Storing the password by modifying the resource was an artifact of the
past anyways, and a login function results in a cleaner system and fixes
the problem at hand.