Keyboard navigation in conversation view.
One problem with the current setting is that j/k is used for scrolling
in the conversation view, but for moving to the next mail in the
maillist view. To normalize this we could be using j/k for moving to the
next conversation instead, and n/p to go to the next/previous mail,
which is also what gmail does.