Make use of interface include directories and link libraries.
Instead of manually specifying the libraries to link against and the
include directories to include we'd much rather have a single target to
link against. find_package already defines the Gpgme target for some
reason, which seems like a waste, but with the lowercase gpgme target we
can work around that problem.