Instead of one model that is really two, use two models based on a common source model.
Both models are supplying the events via a property instead of the model
indexes, but by using models as interface we can change that when
necessary.

This is also the first step towards a month view as we will be abe to
use the multidayeventmodel for the monthview as well.