Fixed encrypted mails
The structure in which we created encrypted mails was not compatible
with what the enigmail plugin expects (because it's apparently a limited
implementation that makes some assumptions), for roundcube and
thunderbird. This is now fixed to generate mail in-line with what the
other clients do:
* The toplevel part is always the signed/encrypted part. Below we then
have e.g. a multipart/alternative if we have attachments.
* The attached key is added as a regular attachment that ends up inside
the encrypted/signed message.