Only resize html page once
The contentSize never shrinks below the actual size and it turns out
that the contentSize can temporarily grow beyond what we would want.
(Seems to be triggreable using reloadAndBypassCache() on the
webengineview).

Only resizing once the page is fully loaded seems to work though.

An alternative approach to determine the size would be to run:

runJavaScript("document.documentElement.scrollHeight", function(result) { console.warn("Got scroll height", result); root.contentHeight = result; });

once the page is loaded, but that requires javascript to be enabled,
and seems to yield the same result as contentSize.