Use a Context singleton to bind account selection together between views.
While I'm not entirely convinced that having a global account selection
is the way to go (we may want a per-window context for instance), it
should cover us for now and is more lightweight to implement compared to
e.g. attached properties.

We're using this so you can switch your current account in any view, and
have it affect the other views as well.