Added support for trimming plaintext emails
In the conversationview we want to avoid showing repetitive information,
especially with the top-posting behaviour of various email clients which
results in very long email's containing the complete past conversation.

Because a solution based on qtextdocument does not seem to work with the
qt quick components, this is now implemented in the partmodel.

For the necessary preprocessing the partmodel now preprocesses the
content as well, and then caches it (which is probably a better idea
anyways than doing it inline in data().

The function that does the trimming simply searches for a delimiter
before which we assume the relevant content is. In case of
inline-replies we assume that delimiter to be at the very top of the
email, and would thus ignore it.

It will probably require some fine-tuning and is therefore only enabled
in experimental mode for now.