Always attempt to unlock all accounts on startup
This avoids the situation that you can interact with an account,
while the account doens't sync because you have never selected it
directly (which is a problem for i.e. the calendar where you don't need
to select the account to see the events).

It's of course still a problem if the accounts don't have a gpg-key and
password stored, because then unlocking won't work.
I suppose we could do a better job in that case of indicating that the
account won't sync, but that seems like a smaller problem for another
time.