Fix mailview loading on startup
Sometimes body.visible ends up set to false on startup,
even though we are in the empty default state.
I was only able to fix the behavior reliably with an explicit state
continaing a when clause, which resets the visible state to where it
should be.

Either a qt bug or we should be doing something different.