Fixed AddresseeEditor sizing
With too many entries on a small screen the Send button at the bottom
would be pushed out of view. This was because the ColumnLayout didn't
size the individual AddresseeEditors correctly.
It seems to me the ColumnLayout doesn't really work as it should, but
with the forced maximumSize this now seems to work as intended.